Remove osprofiler tests
This commit removes running the single osprofiler test on this branch.
This interferes with running unit tests with the --failing flag. Since
we can't backport the stestr migration, which would fix this issue,
because of requirements this just removes that test command. The value
of the osprofiler tests is not as much as being able to use --failing.

Change-Id: I2aaa16867a31a34896c2b9faa44c06b84c2ff17d