Stop assuming initial provider generation is 0
SchedulerReportClient._create_resource_provider used to assume the
generation of a freshly-created resource provider is zero, and store
that value in the provider tree cache.  This violates the opacity of the
generation in the API.

The fix implemented herein is two-pronged.  We exploit the
implementation of the related blueprint, returning the payload of the
POST /resource_providers response if available.  Otherwise (if the
available placement microversion is older), rather than assuming
anything about the new provider, we perform a GET
/resource_providers/{uuid} and return the result (which contains the
generation).

Change-Id: I8e0a37100b0876916db50b08f95f427cb0456616
Closes-Bug: #1746075
Related-Blueprint: generation-from-create-provider