Make nova-cells-v1 run with neutron
This changes the nova-cells-v1 job to run with neutron instead
of nova-network.

SSH validation is already disabled in Tempest for this job.

Since the os-server-external-events API is not plumbed in for
cells v1, we have to disable vif plugging settings in
nova-cells.conf. This means servers will be ACTIVE before
networking is sure to be ready so we can't rely on ssh validation.

We also configure Tempest to say floating IPs aren't available
which skips several scenario tests that wouldn't work with cells v1
anyway due to them trying to ssh into the guest.

Part of blueprint remove-nova-network

Change-Id: I9de6b710baffdffcd1d7ab19897d5776ef27ae7e