Pick next minimum libvirt / QEMU versions for "Stein"
The last time we incremented versions for libvirt and QEMU was during
the "Pike" release (commit: b980df0, 11-Feb-2017); for "Queens" we
didn't do any.  The advertized NEXT_MIN_{LIBVIRT,QEMU} versions during
"Pike" release were set to: libvirt 1.3.1 and QEMU 2.5.0 -- but they
weren't actually bumped for "Queens".  They will be applied for the
"Rocky" release.

Based on the LibvirtDistroSupportMatrix[*], this patch announces a
version increment for libvirt and QEMU that will be applied for the
"Stein" release, with predicted minimum versions setting libvirt to
3.0.0 and QEMU to 2.8.0.

[*] https://wiki.openstack.org/wiki/LibvirtDistroSupportMatrix

Change-Id: I872304f99e3401f9c09d61b5511300ef9c5c7e8d