unquiesce instance after quiesce failure
If the call to compute_rpcapi.quisece_instance() raises an exception,
any uncaught exception will break out of the function
snapshot_volume_backed(). This can leave the instance in frozen state.

This patch adds a blanket Exception catch to the try block and calls
compute_rpcapi.unquiesce_instance() before reraising.

This has been seen in the wild with RPC timeouts, but this is not the
only possible genesis for an unknown error from quiesce_instance.

Change-Id: Idca5998da8bb42b29a8fffdf52b4af3a043c6326
Closes-Bug: #1754360