Report client: Remove version discovery comment
There was a long-standing comment in the report client about performing
version discovery on the placement API.  It was unclear what said
discovery would be used for.  Some thought it should be to allow us to
avoid doing per-API negotiation like:

resp = self.get(..., version=HIGHER)
if resp.status_code == 406:
    resp = self.get(..., version=LOWER)

So a spec [1] and patch [2] were started.  But upon discussion [3] it
was decided that the 406-based negotiation was fine.  So this change set
simply removes the comment.

[1] https://review.openstack.org/549184
[2] https://review.openstack.org/548983
[3] http://eavesdrop.openstack.org/meetings/nova_scheduler/2018/nova_scheduler.2018-03-19-14.00.log.html#l-209

Change-Id: I9bf1491df46d112b523769ab632897c7b9be5a20