Fix message for unexpected external event
During live migration, we get vif plugged events
that we're not waiting for, and if InstanceEvents._events
is empty, pop_instance_event() will log a message about
shutting down, but the compute service isn't actually
shutting down. When it is, ComputeManager.cleanup_host()
calls cancel_all_events() which sets _events to None, so
we should only mention shutting down when _events is None.

Change-Id: Ie868273c4b9629f9b35645df5785304ebba5d64c
Closes-Bug: #1756994