Clarify log in RT._update_usage_from_migration
Now that migration records have UUIDs, this log message
is confusing since it sounds like it's about the migration
record UUID itself, but it's about the instance, and without
context on *what* it's updating, it's more confusing while
tracing a request through the logs. So this clarifies the
actual context of the message and uses the instance_uuid
kwarg for log formatting success.

Change-Id: I5937e6d4c0fb6c3cd6ff2c8b49da35edaba5ef79