Add host field to InstanceActionEvent
This patch adds the host field and obj_make_compatible with
InstanceActionEvent 1.2 version.

We record the event host info in event start method, and do
NOT record the host again in event finish method, because
it's the same event db record which is looked up by the
existing action(via request id) and event name from the
start event, the host of starting and ending is also same.

This would be used by instance action API to show the
``host`` information of the action event occurs.

Part of blueprint: add-host-to-instance-action-events

Change-Id: I356e7eb6b18837cf868bade7babaed96e5828688