Noauth should also use request_id from compute_req_id.py
When using noauth midlleware, the context is generated here:
https://github.com/openstack/nova/blob/ef6f4e4c8ec82e2c9f9988fe2e04591ee01220e6/nova/api/openstack/auth.py#L56
as the openstack.request_id from req.environ, previously created in:
https://github.com/openstack/oslo.middleware/blob/master/oslo_middleware/request_id.py#L57
was not included in the initialize of the context obj,
oslo_context.context.Context.__init__ will generate a new request_id for
context obj.

As the req.environ['openstack.request_id'] will later be returned to the
user as the 'x-openstack-request-id' field in the response header, users
may use this for searching etc. But the Nova workflow will all use the
context.request_id, which will lead to an inconsistency problem.

Change-Id: I047b5f66b33ac89b2b5e9170c12bc87d76d0946b
Closes-Bug: #1758031