Add the version description for InstanceActionEventList
This change is from [1], but now, this version up is not
necessary to merge, so I submit the description in here.

I know this is a trivial change, but I just want to explain
that I hope somebody want to do version up for
InstanceActionEventList, and will no need to look the
commit history[2] again.

[1] https://review.openstack.org/#/c/555146/2/nova/objects/instance_action.py@227
[2] v1.1 version change If9189c57290dcb19a0efd94bcfa02f9c4eef53d1
[3] v1.1 code change Ibedddd606b9d25ffbbb2b1b5358cf0aa4741083f

Change-Id: Idb3196b4f0ece8990267b02fd24f7a3b747ef22a