Avoid unnecessary port update during live migration
The ComputeManager.pre_live_migration method, which runs on the
destination host during live migration, calls setup_networks_on_host
with the destination host which causes the "migrating_to" attribute
to get set to the destination host in the port binding profile. This
is to initiate setting up floating IP information on the destination
host for DVR.

The ComputeManager.post_live_migration_at_destination method calls the
same method from the destination host, so at this point the "migrating_to"
attribute is already set to the destination host in the port binding
profile from pre_live_migration and we make unnecessary port update calls
to Neutron for each port attached to the instance.

This change simply checks to see if the "migrating_to" attribute is
already set to the given host and if so, skips the port update call.

Change-Id: I8b2ffd94c5ff0881d0df4efdde9be938888c16d7
Closes-Bug: #1758453