Delete port bindings in setup_networks_on_host if teardown=True
This change adds some logic to the neutronv2 API implementation
of setup_networks_on_host such that if teardown=True and a host
is provided (which is not the current instance.host), port
bindings, for the instance in scope, for that host are deleted.

If the port bindings don't exist, then Neutron will return a 404
which we ignore.

This will be used later in _post_live_migration and
_rollback_live_migration flows. In the post case, the live
migration was successful so the inactive port bindings on the
source host will be deleted. In the rollback case, the live
migration failed so the inactive port bindings on the destination
host will be deleted.

Part of blueprint neutron-new-port-binding-api

Change-Id: I588d6845c69a8cb751dc17b615884985a5c40155