Implement migrate_instance_start method for neutron
This implements the migrate_instance_start() method for the
neutronv2 API code such that if the neutron port bindings-extended
API extension is available and the ports attached to the instance
have bindings to the destination host, those bindings are activated.

This will make the source compute host port bindings inactive and
also update the port binding:host_id attribute.

This is part of a larger series of changes for live migration where
ComputeManager._post_live_migration calls migrate_instance_start
to activate the bindings on the dest host.

None of this "turns on" until the final patch in the series where
conductor creates the destination host port bindings.

Part of blueprint neutron-new-port-binding-api

Change-Id: I510e0707b63da98af143448dd46b771f3ef74624