Add host to API and Conductor
The instance action event is recorded in API/Conductor/Compute using
EventReporter, in order to help admins to find the host of events
occurring we also need pass the host info to EventReporter.

In compute, we already have a self.host on the
nova.compute.manager.ComputeManager, but API/Conductor doesn't contain
this.

In order to get host conveniently, this patch add host to API and
Conductor, and set it to CONF.host when api and conductor initialize.

Part of blueprint: add-host-to-instance-action-events

Change-Id: Idf57fb5fbc611abb83943bd7e36d3cebf03b3977