Unit test framework: common FakeResponse
Faking out requests.Response as a mock return value can be awkward,
particularly due to how a real Response responds to a boolean check
(False for status codes >=400).  Attempting to use a real Response is
also awkward when it needs to have a payload, due to indirection and
encoding.

This change set introduces a module, nova.tests.unit.fake_requests,
intended to hold fakes for things found in the requests module.  For
now, it holds FakeResponse, which provides a convenient way to create a
Response-like that can be used as a mock return value, which responds
like a real Response to most operations on it.

We swap this into various test modules that were previously using other
mechanisms to fake Response instances.

Change-Id: Iaeced55c442afdcb277de97829f52fe3c02a380a