doc: Upgrade placement first
It has long been *recommended* that the placement service be upgraded
before nova services, but scenarios where it isn't have never really
been explored or tested.  Following (yet another) discussion on the
topic [1] it was decided that we are best served by *requiring* that
placement be updated first.  This simplifies our test/support matrix,
and the code we write to consume placement.  It also makes the operator
experience cleaner by giving them a clear and unambiguous script to
follow.

This change set rewords the upgrade document accordingly.

[1] http://eavesdrop.openstack.org/irclogs/%23openstack-nova/%23openstack-nova.2018-03-26.log.html#t2018-03-26T17:35:11

Change-Id: I97215e94efdd8c05045872fb9ba7d2089dc6efb8