Start untangling network utilities.
The nova-net code contains a set of network utilities that are more
generally useful than nova-net, and that are called from a variety
of places in the nova code. Start separating those out into a utility
module so its clearer that they're not in fact part of nova-net and
should survive past the removal of nova-net.

We start with device_exists, because it is the first one I noticed.

Change-Id: Ieac8621217c42f8b2d08dbc58c2025262f325e1e
blueprint: remove-nova-network