Remove mox in tests/unit/test_utils.py
Replace mox with mock in nova/tests/unit/test_utils.py.
The 'test_supports_direct_io' method is
in the 'GetEndpointTestCase' class currently.
But it should be in an isolated test class.
So add the test class and move the method into it.

Then split the method into methods for each test case
to improve readability.

Change-Id: Id2350529b3322dfa8f7c13ac8e5f85aaf3041082
Implements: blueprint mox-removal
Closes-Bug: #1760017