Wait for network-vif-plugged before starting live migration
This adds a new config option which is read on the source host
during pre_live_migration which can be used to determine
if it should wait for a "network-vif-plugged" event due to VIFs
being plugged on the destination host. This helps us to
avoid the guest transfer at all if vif plugging failed on the dest
host, which we just wouldn't find out until post live migration
and then we have to rollback.

The option is disabled by default for backward compatibility and
also because certain networking backends, like OpenDaylight, are
known to not send network-vif-plugged events unless the port host
binding information changes, which for live migration doesn't happen
until after the guest is transferred to the destination host.

Related to blueprint neutron-new-port-binding-api
Related-Bug: #1786346

NOTE(danms): Stable-only changes to this patch from master include
removing the RPC-related communication from the destination
to the source node. As such, the new option is read on the source
node so the conf option help and release note are updated. This is
OK before Rocky since we don't claim support to live migrate
between different networking backends (vif types), so operators
would need to set the option universally, or at least have host
aggregates in place if they are using different network types.

Conflicts:
        nova/conf/compute.py
        nova/tests/unit/objects/test_migrate_data.py

Change-Id: I0f3ab6604d8b79bdb75cf67571e359cfecc039d8
(cherry picked from commit 5aadff75c3ac4f2019838600df6580481a96db0f)